requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
accelerate>=0.20.0